+++ ta1/NodeFactNum.java
@@ -1,3 +1,6 @@
+/**
+ * Class to pull the value of a number from our environment
+ */
 public class NodeFactNum extends NodeFact {
 
 	private String num;
@@ -5,11 +8,16 @@ public class NodeFactNum extends NodeFact {
 	public NodeFactNum(String num) {
 		this.num=num;
 	}
-
-	public int eval(Environment env) throws EvalException {
-		return Integer.parseInt(num);
+/**
+ * return the value of the double stored at this position in the string
+ */
+	public double eval(Environment env) throws EvalException {
+		return Double.parseDouble(num);
 	}
 
+	/**
+	 * return value of number to generate in C code
+	 */
 	public String code() { return num; }
 
 }
